flask_app/requirements.txt
appdirs==1.4.4
attrs==19.3.0
audioread==2.1.8
autopep8==1.5.3
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Uploads==0.2.1
future==0.18.2
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mpmath==1.1.0
my-voice-analysis==0.7
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.0.3
numba==0.51.0
numpy==1.19.1
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
playsound==1.2.2
pooch==1.1.1
praat-parselmouth==0.3.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydub==0.24.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
python-speech-features==0.6
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sympy==1.6.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1